// frontend/lyric-sync-app/src/api/models/GeniusHitDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ChordKTV API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GeniusResultDto } from './GeniusResultDto';
import {
    GeniusResultDtoFromJSON,
    GeniusResultDtoFromJSONTyped,
    GeniusResultDtoToJSON,
    GeniusResultDtoToJSONTyped,
} from './GeniusResultDto';

/**
 * 
 * @export
 * @interface GeniusHitDto
 */
export interface GeniusHitDto {
    /**
     * 
     * @type {GeniusResultDto}
     * @memberof GeniusHitDto
     */
    result: GeniusResultDto;
}

/**
 * Check if a given object implements the GeniusHitDto interface.
 */
export function instanceOfGeniusHitDto(value: object): value is GeniusHitDto {
    if (!('result' in value) || value['result'] === undefined) return false;
    return true;
}

export function GeniusHitDtoFromJSON(json: any): GeniusHitDto {
    return GeniusHitDtoFromJSONTyped(json, false);
}

export function GeniusHitDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): GeniusHitDto {
    if (json == null) {
        return json;
    }
    return {
        
        'result': GeniusResultDtoFromJSON(json['result']),
    };
}

export function GeniusHitDtoToJSON(json: any): GeniusHitDto {
    return GeniusHitDtoToJSONTyped(json, false);
}

export function GeniusHitDtoToJSONTyped(value?: GeniusHitDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'result': GeniusResultDtoToJSON(value['result']),
    };
}